refacotr(dsg): moving common logic
