+++ packages/amplication-data-service-generator/src/server/resource/dto/to-many/create-create-nested-many-without-input.ts
@@ -1,36 +1,25 @@
 import { builders, namedTypes } from "ast-types";
 import { pascalCase } from "pascal-case";
-import { Entity, EntityField } from "../../../../types";
-import { classDeclaration, NamedClassDeclaration } from "../../../../util/ast";
+import { Entity } from "../../../../types";
+import { NamedClassDeclaration } from "../../../../util/ast";
 import { isToManyRelationField } from "../../../../util/field";
-import { INPUT_TYPE_DECORATOR } from "../nestjs-graphql.util";
-import { createCreateNestedManyProperties } from "./connectOrCreateClassBody";
+import { createNestedInputDTO } from "./create-nested";
 
 export function createCreateNestedManyDTOs(
   entity: Entity
 ): NamedClassDeclaration[] {
   const toManyFields = entity.fields.filter(isToManyRelationField);
   const createNestedManyWithoutInputDtos = toManyFields.map((field) =>
-    createCreateNestedManyDTO(entity, field)
+    createNestedInputDTO(
+      createCreateNestedManyWithoutInputID(
+        entity.pluralDisplayName,
+        field.properties.relatedEntity.name
+      ),
+      entity,
+      field
+    )
   );
   return createNestedManyWithoutInputDtos;
-}
-
-export function createCreateNestedManyDTO(
-  entity: Entity,
-  toManyField: EntityField
-): NamedClassDeclaration {
-  const properties = createCreateNestedManyProperties(toManyField, entity);
-  const decorators = properties.length ? [INPUT_TYPE_DECORATOR] : [];
-  return classDeclaration(
-    createCreateNestedManyWithoutInputID(
-      entity.pluralDisplayName,
-      toManyField.properties.relatedEntity.name
-    ),
-    builders.classBody(properties),
-    null,
-    decorators
-  ) as NamedClassDeclaration;
 }
 
 export function createCreateNestedManyWithoutInputID(
